clean (FileUploader.cs): split it up into partial files for the sake of readability
