+++ Laerdal.McuMgr/Shared/FileUploading/Contracts/IFileUploaderCleanupable.cs
@@ -2,6 +2,6 @@ namespace Laerdal.McuMgr.FileUploading.Contracts
 {
     public interface IFileUploaderCleanupable
     {
-        void CleanupResourcesOfLastUpload();
+        void TryCleanupResourcesOfLastUpload();
     }
 }
