+++ Laerdal.McuMgr/Shared/FileUploading/Contracts/Native/INativeFileUploaderCleanupProxy.cs
@@ -2,6 +2,6 @@ namespace Laerdal.McuMgr.FileUploading.Contracts.Native
 {
     public interface INativeFileUploaderCleanupProxy
     {
-        void CleanupResourcesOfLastUpload();
+        void TryCleanupResourcesOfLastUpload();
     }
 }
