+++ Laerdal.McuMgr/Shared/FileUploading/FileUploader.Cleanupable.TryCleanupResourcesOfLastUpload.cs
@@ -1,0 +1,7 @@
+namespace Laerdal.McuMgr.FileUploading
+{
+    public partial class FileUploader
+    {
+        public void TryCleanupResourcesOfLastUpload() => NativeFileUploaderProxy?.TryCleanupResourcesOfLastUpload();
+    }
+}
