+++ Laerdal.McuMgr/Shared/FileUploading/FileUploader.Commandable.TryDisconnect.cs
@@ -1,0 +1,7 @@
+namespace Laerdal.McuMgr.FileUploading
+{
+    public partial class FileUploader
+    {
+        public bool TryDisconnect() => NativeFileUploaderProxy?.TryDisconnect() ?? false;
+    }
+}
