+++ Laerdal.McuMgr/Shared/FileUploading/FileUploader.Commandable.TryInvalidateCachedInfrastructure.cs
@@ -1,0 +1,7 @@
+namespace Laerdal.McuMgr.FileUploading
+{
+    public partial class FileUploader
+    {
+        public bool TryInvalidateCachedInfrastructure() => NativeFileUploaderProxy?.TryInvalidateCachedInfrastructure() ?? false;
+    }
+}
