+++ Laerdal.McuMgr/Shared/FileUploading/FileUploader.Commandable.TrySetBluetoothDevice.cs
@@ -1,0 +1,7 @@
+namespace Laerdal.McuMgr.FileUploading
+{
+    public partial class FileUploader
+    {
+        public bool TrySetBluetoothDevice(object bluetoothDevice) => NativeFileUploaderProxy?.TrySetBluetoothDevice(bluetoothDevice) ?? false;
+    }
+}
