+++ Laerdal.McuMgr/Shared/FileUploading/FileUploader.Commandable.TrySetContext.cs
@@ -1,0 +1,7 @@
+namespace Laerdal.McuMgr.FileUploading
+{
+    public partial class FileUploader
+    {
+        public bool TrySetContext(object context) => NativeFileUploaderProxy?.TrySetContext(context) ?? false;
+    }
+}
