+++ Laerdal.McuMgr/Shared/FileUploading/FileUploader.Commandable.cs
@@ -1,0 +1,9 @@
+// ReSharper disable UnusedType.Global
+// ReSharper disable RedundantExtendsListEntry
+
+namespace Laerdal.McuMgr.FileUploading
+{
+    public partial class FileUploader
+    {
+    }
+}
