+++ Laerdal.McuMgr/Shared/FirmwareInstallation/Contracts/Native/INativeFirmwareInstallerCleanupableProxy.cs
@@ -2,6 +2,6 @@ namespace Laerdal.McuMgr.FirmwareInstallation.Contracts.Native
 {
     internal interface INativeFirmwareInstallerCleanupableProxy
     {
-        void CleanupResourcesOfLastInstallation();
+        void TryCleanupResourcesOfLastInstallation();
     }
 }
